backend/requirements/base.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
appdirs==1.4.4
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
Bottleneck==1.6.0
bs4==0.0.2
build==1.4.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
curl_cffi==0.13.0
dnspython==2.8.0
empyrical-reloaded==0.5.12
eod==0.2.1
feedparser==6.0.12
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.1.0
gast==0.7.0
google-auth==2.47.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
greenlet==3.3.1
grpcio==1.76.0
h5py==3.15.1
html5lib==1.1
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iso3166==2.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.13.1
korean-lunar-calendar==0.3.1
libclang==18.1.1
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
more-itertools==10.8.0
multidict==6.7.0
multipledispatch==1.0.0
multitasking==0.0.12
mysql==0.0.3
mysql-connector-python==9.5.0
mysqlclient==2.2.7
namex==0.1.0
nulltype==2.3.1
numpy==2.4.1
nvidia-nccl-cu12==2.29.2
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pandas-datareader==0.10.0
patsy==1.0.2
peewee==3.17.3
pillow==12.1.0
plaid-python==38.0.0
platformdirs==4.5.1
polygon-api-client==1.16.3
premailer==3.10.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.16.0
pyOpenSSL==25.3.0
pyparsing==3.3.2
pyppeteer==0.0.25
pyproject_hooks==1.2.0
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
setuptools==80.10.1
setuptools-scm==9.2.2
sgmllib3k==1.0.0
six==1.17.0
sklearn-compat==0.1.5
soupsieve==2.8.3
SQLAlchemy==2.0.46
stripe==14.2.0
ta==0.11.0
TA-Lib==0.6.8
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
threadpoolctl==3.6.0
tomli==2.4.0
tqdm==4.67.1
twilio==9.10.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
webencodings==0.5.1
websockets==16.0
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.0.1
xgboost==3.1.3
yarl==1.22.0
yfinance==1.0
zipp==3.23.0
